+++ NEWS.md
@@ -1,5 +1,6 @@
 # Unreleased version
 
+- Add `delay` argument to show the spinner after a short delay (#95)
 - Add `width` argument that can be used in rare cases where the spinner has no inherent width (#85)
 
 # shinycssloaders 1.1.0 (2024-07-30)